ecoScripts/aurora/Image/snow-table-parser/requirements.txt
pyyaml
requests
bs4
pykafka